+++ lib/src/result_of.dart
@@ -29,11 +29,19 @@ class ResultOf<T> extends Result {
   // ignore: prefer_constructors_over_static_methods
   /// Create fail `Result` with reason of fail
   /// ```dart
-  /// ResultOf.fail<MyObject>('fail reason');
+  /// ResultOf.fail<MyObject>(ResultError('fail reason'));
   /// ```
   static ResultOf<T> fail<T>(ResultError error) {
     return ResultOf<T>(isSuccess: false, value: null, error: error);
   }
+
+  ///
+  static ResultOf<T> failMessage<T>(String message) =>
+      ResultOf<T>(isSuccess: false, value: null, error: ResultError(message));
+
+  ///
+  static ResultOf<T> exception<T>(Exception exception) => ResultOf<T>(
+      isSuccess: false, value: null, error: ResultException(exception));
 
   /// <summary>
   /// Convert result with value to result with another value. Use valueConverter